paging fixed

+++ src/kernel/utils.c
@@ -1,7 +1,7 @@
 #include "types.h"
 #include "utils.h"
 
-void memcpy(char *src, char *dest, uint64 size)
+void memcpy(char *src, char *dest, size_t size)
 {
     while (size--) {
         dest[size] = src[size];
